Clean-up/bug fixes (#718)
* updates

* fixes

* testing fix
