+++ bittensor/_neuron/text/advanced_server/nucleus_impl.py
@@ -46,7 +46,8 @@ class server(torch.nn.Module):
                     Custom function that maps between tokenizers (defaults to self.remapping_token)
         """
         super(server, self).__init__()
-        if config == None: config = server.config()
+        if config == None: 
+            raise Exception('No Config Given: Config Required')
         self.config = config;print(config)
         
         #setting up pretrained model
@@ -198,38 +199,3 @@ class server(torch.nn.Module):
 
         except Exception as e:
             logger.warning('No saved model found with error: {}', e)
-
-    @staticmethod
-    def config ():
-        parser = argparse.ArgumentParser()
-        parser.add_argument('--config', type=str, help='If set, defaults are overridden by passed file.')
-        parser.add_argument('--neuron.learning_rate', type=float, help='Training initial learning rate.', default=0.01)
-        parser.add_argument('--neuron.momentum', type=float, help='optimizer momentum.', default=0.8)
-        parser.add_argument('--neuron.clip_gradients', type=float, help='Implement gradient clipping to avoid exploding loss on smaller architectures.', default=1.0)
-        parser.add_argument('--neuron.device', type=str, help='miner default training device cpu/cuda', default=("cuda" if torch.cuda.is_available() else "cpu"))
-        parser.add_argument('--neuron.model_name', type=str, help='pretrained model from hugging face',default='gpt2')
-        parser.add_argument('--neuron.pretrained', action='store_false', help='if the model should be pretrained',default=True)
-        parser.add_argument('--neuron.padding', action='store_false', help='To pad out final dimensions',default=True)
-        parser.add_argument('--neuron.interpolate', action='store_false', help='To interpolate between sentence length',default=True)
-        parser.add_argument('--neuron.inter_degree', type=str, help='Interpolate algorithm (nearest | linear | bilinear | bicubic | trilinear | area)', default='nearest')
-        parser.add_argument('--neuron.name', type=str, help='Trials for this miner go in miner.root / (wallet_cold - wallet_hot) / miner.name ', default='advanced_server')
-        parser.add_argument('--neuron.checking', action='store_false', help='To check if server settings are correct',default=True)
-        parser.add_argument('--neuron.restart', action='store_true', help='If set, train the neuron from the beginning', default=False)
-        parser.add_argument('--neuron.blacklist.stake.forward', type=float, help='Amount of stake (tao) in order not to get blacklisted for forward requests', default=10)
-        parser.add_argument('--neuron.blacklist.stake.backward', type=float, help='Amount of stake (tao) in order not to get blacklisted for backward requests', default=100)
-        parser.add_argument('--neuron.blacklist_allow_non_registered', action='store_true', help='''If true, black lists non-registered peers''', default=True)
-        parser.add_argument('--neuron.metagraph_sync', type=float, help='how often to sync the metagraph', default=100000)
-        parser.add_argument('--neuron.blocks_per_set_weights', type=float, help='how often to sync set weights', default=100)
-        parser.add_argument('--neuron.blocks_per_epoch', type=int, help='Blocks per epoch', default=2)
-        parser.add_argument('--neuron.blacklist.time', type=int, help='how often a peer can query you (seconds) ', default=0)
-
-        bittensor.wallet.add_args( parser )
-        bittensor.axon.add_args( parser )
-        bittensor.subtensor.add_args( parser )
-        bittensor.logging.add_args( parser )
-        bittensor.wandb.add_args(parser)
-        bittensor.prioritythreadpool.add_args( parser )
-        bittensor.dataset.add_args( parser )
-        bittensor.metagraph.add_args( parser )
-        return bittensor.config( parser )
-
