+++ bittensor/_neuron/text/template_server/__init__.py
@@ -24,6 +24,7 @@ Example:
 
 import bittensor
 import os
+import argparse
 
 from .nucleus_impl import server
 from .run import serve
@@ -60,7 +61,7 @@ class neuron:
         metagraph: 'bittensor.metagraph' = None,
 
     ):
-        if config == None: config = server.config()
+        if config == None: config = neuron.config()
         config = config; 
         self.check_config( config )
         bittensor.logging (
@@ -86,11 +87,6 @@ class neuron:
             metagraph= self.metagraph,
         )
 
-
-    @classmethod
-    def config(cls):
-        return server.config()
-
     @staticmethod
     def check_config( config: 'bittensor.Config' ):
         r""" Checks/validates the config namespace object.
@@ -106,3 +102,38 @@ class neuron:
         config.neuron.full_path = os.path.expanduser(full_path)
         if not os.path.exists(config.neuron.full_path):
             os.makedirs(config.neuron.full_path)
+
+    @staticmethod
+    def config ():
+        parser = argparse.ArgumentParser()
+        parser.add_argument('--config', type=str, help='If set, defaults are overridden by passed file.')
+        parser.add_argument('--neuron.learning_rate', type=float, help='Training initial learning rate.', default=0.01)
+        parser.add_argument('--neuron.momentum', type=float, help='optimizer momentum.', default=0.8)
+        parser.add_argument('--neuron.clip_gradients', type=float, help='Implement gradient clipping to avoid exploding loss on smaller architectures.', default=1.0)
+        parser.add_argument('--neuron.device', type=str, help='miner default training device cpu/cuda', default=("cuda" if torch.cuda.is_available() else "cpu"))
+        parser.add_argument('--neuron.model_name', type=str, help='pretrained model from hugging face',default='gpt2')
+        parser.add_argument('--neuron.pretrained', action='store_false', help='if the model should be pretrained',default=True)
+        parser.add_argument('--neuron.padding', action='store_false', help='To pad out final dimensions',default=True)
+        parser.add_argument('--neuron.interpolate', action='store_false', help='To interpolate between sentence length',default=True)
+        parser.add_argument('--neuron.inter_degree', type=str, help='Interpolate algorithm (nearest | linear | bilinear | bicubic | trilinear | area)', default='nearest')
+        parser.add_argument('--neuron.name', type=str, help='Trials for this miner go in miner.root / (wallet_cold - wallet_hot) / miner.name ', default='advanced_server')
+        parser.add_argument('--neuron.checking', action='store_false', help='To check if server settings are correct',default=True)
+        parser.add_argument('--neuron.restart', action='store_true', help='If True, train the neuron from the beginning', default=False)
+        parser.add_argument('--neuron.blacklist.stake', type=float, help='Amount of stake (tao) in order not to get blacklisted', default=0)
+        parser.add_argument('--neuron.blocks_per_epoch', type=int, help='Blocks per epoch', default=10)
+        parser.add_argument('--neuron.blacklist.time', type=int, help='how often a peer can query you (seconds) ', default=1)
+        parser.add_argument('--neuron.training',  action='store_true', help='if the model should be training (increases memory load)', default=False)
+        parser.add_argument('--neuron.autocast',  action='store_true', help='(experimental) autocasts the model to float16. Must require cuda', default=False)
+        parser.add_argument('--neuron.blocks_per_set_weights', type=float, help='how often to set weights', default=100)
+        parser.add_argument('--neuron.metagraph_sync', type=float, help='how often to sync the metagraph', default=100000)
+
+
+        bittensor.wallet.add_args( parser )
+        bittensor.axon.add_args( parser )
+        bittensor.subtensor.add_args( parser )
+        bittensor.logging.add_args( parser )
+        bittensor.wandb.add_args(parser)
+        bittensor.prioritythreadpool.add_args( parser )
+        bittensor.dataset.add_args( parser )
+        bittensor.metagraph.add_args( parser )
+        return bittensor.config( parser )
